+++ app/src/main/java/de/markusfisch/android/shadereditor/widget/ShaderEditor.java
@@ -3,9 +3,11 @@ package de.markusfisch.android.shadereditor.widget;
 import android.content.Context;
 import android.graphics.Canvas;
 import android.graphics.Paint;
+import android.os.Build;
 import android.os.Handler;
 import android.text.Editable;
 import android.text.InputFilter;
+import android.text.InputType;
 import android.text.Layout;
 import android.text.Spannable;
 import android.text.SpannableStringBuilder;
@@ -17,16 +19,24 @@ import android.text.style.ForegroundColorSpan;
 import android.text.style.LineHeightSpan;
 import android.text.style.ReplacementSpan;
 import android.util.AttributeSet;
+import android.util.Log;
 import android.view.KeyEvent;
-import android.view.View;
+import android.view.inputmethod.EditorInfo;
+import android.view.inputmethod.InputConnection;
 
 import androidx.annotation.ColorInt;
 import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
+import androidx.annotation.RequiresApi;
 import androidx.core.content.ContextCompat;
 
 import java.util.ArrayList;
 import java.util.List;
+import java.util.concurrent.Callable;
+import java.util.concurrent.ExecutionException;
+import java.util.concurrent.ExecutorService;
+import java.util.concurrent.Executors;
+import java.util.concurrent.FutureTask;
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
 
@@ -42,6 +52,11 @@ public class ShaderEditor extends LineNumberEditText {
 		void onTextChanged(String text);
 	}
 
+	@FunctionalInterface
+	public interface CodeCompletionListener {
+		void onCodeCompletions(@NonNull List<String> completions, int position);
+	}
+
 	private static final Pattern PATTERN_TRAILING_WHITE_SPACE = Pattern.compile(
 			"[\\t ]+$",
 			Pattern.MULTILINE);
@@ -55,7 +70,17 @@ public class ShaderEditor extends LineNumberEditText {
 	private static final Pattern PATTERN_MAIN = Pattern.compile(
 			".*void\\s+main\\s*\\(.*");
 	private static final Pattern PATTERN_NO_BREAK_SPACE = Pattern.compile(
-			"[\\xA0]");
+			"\\xA0");
+	private static final ArrayList<String> DEFAULT_COMPLETIONS = new ArrayList<>();
+
+	static {
+		DEFAULT_COMPLETIONS.add("{");
+		DEFAULT_COMPLETIONS.add("}");
+		DEFAULT_COMPLETIONS.add("(");
+		DEFAULT_COMPLETIONS.add(")");
+		DEFAULT_COMPLETIONS.add("[");
+		DEFAULT_COMPLETIONS.add("]");
+	}
 
 	private final Handler updateHandler = new Handler();
 	private final Runnable updateRunnable = new Runnable() {
@@ -63,16 +88,21 @@ public class ShaderEditor extends LineNumberEditText {
 		public void run() {
 			Editable e = getText();
 
-			if (onTextChangedListener != null) {
-				onTextChangedListener.onTextChanged(e.toString());
+			if (e != null) {
+				if (onTextChangedListener != null) {
+					onTextChangedListener.onTextChanged(e.toString());
+				}
+
+				highlightWithoutChange(e);
 			}
 
-			highlightWithoutChange(e);
 		}
 	};
 	private final int[] colors = new int[Highlight.values().length];
 
 	private OnTextChangedListener onTextChangedListener;
+	@Nullable
+	private CodeCompletionListener codeCompletionListener;
 	private int updateDelay = 1000;
 	private int errorLine = 0;
 	private boolean dirty = false;
@@ -82,6 +112,10 @@ public class ShaderEditor extends LineNumberEditText {
 	private int tabWidthInCharacters = 0;
 	private int tabWidth = 0;
 	private List<Token> tokens = new ArrayList<>();
+	private int revision = 0;
+	private final TokenListUpdater tokenListUpdater =
+			new TokenListUpdater(this::provideCompletions);
+	private boolean editing = false;
 
 	public ShaderEditor(Context context) {
 		super(context);
@@ -95,6 +129,10 @@ public class ShaderEditor extends LineNumberEditText {
 
 	public void setOnTextChangedListener(OnTextChangedListener listener) {
 		onTextChangedListener = listener;
+	}
+
+	public void setOnCompletionsListener(@NonNull CodeCompletionListener listener) {
+		codeCompletionListener = listener;
 	}
 
 	public void setUpdateDelay(int ms) {
@@ -119,7 +157,10 @@ public class ShaderEditor extends LineNumberEditText {
 	}
 
 	public void updateHighlighting() {
-		highlightWithoutChange(getText());
+		Editable text = getText();
+		if (text != null) {
+			highlightWithoutChange(text);
+		}
 	}
 
 	private void clearError(@Nullable Spannable e) {
@@ -165,11 +206,14 @@ public class ShaderEditor extends LineNumberEditText {
 		dirty = false;
 
 		modified = false;
+
+		tokenListUpdater.update(text, ++revision); // `setText` can't be overridden
 		setText(highlight(new SpannableStringBuilder(text), true));
 		modified = true;
 
-		if (onTextChangedListener != null) {
-			onTextChangedListener.onTextChanged(getText().toString());
+		Editable e = getText();
+		if (e != null && onTextChangedListener != null) {
+			onTextChangedListener.onTextChanged(e.toString());
 		}
 	}
 
@@ -179,16 +223,18 @@ public class ShaderEditor extends LineNumberEditText {
 				.replaceAll("");
 	}
 
-	public void insertTab() {
+	public void insert(@NonNull CharSequence text) {
 		int start = getSelectionStart();
 		int end = getSelectionEnd();
-
-		getText().replace(
-				Math.min(start, end),
+		Editable e = getText();
+		if (e == null) {
+			return;
+		}
+		e.replace(Math.min(start, end),
 				Math.max(start, end),
-				"\t",
+				text,
 				0,
-				1);
+				text.length());
 	}
 
 	public void addUniform(String statement) {
@@ -226,7 +272,46 @@ public class ShaderEditor extends LineNumberEditText {
 			++start;
 		}
 
-		e.insert(start, statement);
+		if (e != null) {
+			e.insert(start, statement);
+		}
+	}
+
+	/**
+	 * Define functions required for AutoFill API
+	 */
+	@RequiresApi(api = Build.VERSION_CODES.O)
+	@Override
+	public int getAutofillType() {
+		return AUTOFILL_TYPE_NONE;
+	}
+
+	@Nullable
+	@Override
+	public InputConnection onCreateInputConnection(@NonNull EditorInfo outAttrs) {
+		InputConnection connection = super.onCreateInputConnection(outAttrs);
+		if (ShaderEditorApp.preferences.hideNativeSuggestions()) {
+			outAttrs.inputType = InputType.TYPE_NULL;
+		}
+		return connection;
+	}
+
+	@Override
+	protected void onSelectionChanged(int selStart, int selEnd) {
+		super.onSelectionChanged(selStart, selEnd);
+		Editable text = getText();
+		CodeCompletionListener listener = codeCompletionListener;
+		if (text == null || listener == null) {
+			return;
+		}
+		int start = getSelectionStart();
+		int end = getSelectionEnd();
+		if (start != end) {
+			return;
+		}
+		if (!editing && tokenListUpdater.isDone()) {
+			provideCompletions(tokens, text);
+		}
 	}
 
 	private void removeUniform(Editable e, String statement) {
@@ -297,6 +382,7 @@ public class ShaderEditor extends LineNumberEditText {
 					int start,
 					int count,
 					int after) {
+				editing = true;
 			}
 
 			@Override
@@ -310,10 +396,13 @@ public class ShaderEditor extends LineNumberEditText {
 				}
 
 				if (!modified) {
+					editing = false;
 					return;
 				}
 
+				editing = false;
 				dirty = true;
+				tokenListUpdater.update(e, ++revision);
 				updateHandler.postDelayed(updateRunnable, updateDelay);
 			}
 		});
@@ -322,19 +411,16 @@ public class ShaderEditor extends LineNumberEditText {
 		setUpdateDelay(ShaderEditorApp.preferences.getUpdateDelay());
 		setTabWidth(ShaderEditorApp.preferences.getTabWidth());
 
-		setOnKeyListener(new OnKeyListener() {
-			@Override
-			public boolean onKey(View v, int keyCode, KeyEvent event) {
-				if (ShaderEditorApp.preferences.useTabForIndent() &&
-						event.getAction() == KeyEvent.ACTION_DOWN &&
-						keyCode == KeyEvent.KEYCODE_TAB) {
-					// Insert a tab character instead of doing focus
-					// navigation.
-					insertTab();
-					return true;
-				}
-				return false;
+		setOnKeyListener((v, keyCode, event) -> {
+			if (ShaderEditorApp.preferences.useTabForIndent() &&
+					event.getAction() == KeyEvent.ACTION_DOWN &&
+					keyCode == KeyEvent.KEYCODE_TAB) {
+				// Insert a tab character instead of doing focus
+				// navigation.
+				insert("\t");
+				return true;
 			}
+			return false;
 		});
 	}
 
@@ -352,13 +438,13 @@ public class ShaderEditor extends LineNumberEditText {
 		updateHandler.removeCallbacks(updateRunnable);
 	}
 
-	private void highlightWithoutChange(Editable e) {
+	private void highlightWithoutChange(@NonNull Editable e) {
 		modified = false;
 		highlight(e, false);
 		modified = true;
 	}
 
-	private Editable highlight(Editable e, boolean complete) {
+	private Editable highlight(@NonNull Editable e, boolean complete) {
 		int length = e.length();
 
 		clearError(e);
@@ -380,40 +466,57 @@ public class ShaderEditor extends LineNumberEditText {
 			clearSpans(e, 0, length, ForegroundColorSpan.class);
 			return e;
 		}
-
-		Lexer lexer = new Lexer(e.toString());
-		List<Token> oldTokens = tokens;
-		tokens = new ArrayList<>();
-		for (Token token : lexer) {
-			tokens.add(token);
-		}
+		List<Token> newTokens = tokenListUpdater.ensureUpdated(e, revision);
 
 		if (complete) {
 			clearSpans(e, 0, length, ForegroundColorSpan.class);
-			for (Token token : tokens) {
+			for (Token token : newTokens) {
 				@ColorInt int color = colors[Highlight.from(token.type()).ordinal()];
 				if (color != textColor) {
 					e.setSpan(
 							new ForegroundColorSpan(color),
-							token.startOffset(), token.endOffset(), Spanned.SPAN_EXCLUSIVE_EXCLUSIVE);
+							token.startOffset(), token.endOffset(),
+							Spanned.SPAN_EXCLUSIVE_EXCLUSIVE);
 				}
 			}
 		} else {
-			Lexer.Diff diff = Lexer.diff(oldTokens, tokens);
+			Lexer.Diff diff = Lexer.diff(tokens, newTokens);
 			if (diff.start <= diff.deleteEnd) {
-				int startOffset = tokens.get(diff.start).startOffset();
-				int endOffset = tokens.get(diff.insertEnd).endOffset();
+				int startOffset = newTokens.get(diff.start).startOffset();
+				int endOffset = newTokens.get(diff.insertEnd).endOffset();
 				clearSpans(e, startOffset, endOffset, ForegroundColorSpan.class);
 			}
 			for (int i = diff.start; i <= diff.insertEnd; ++i) {
-				Token token = tokens.get(i);
+				Token token = newTokens.get(i);
 				e.setSpan(
 						new ForegroundColorSpan(colors[Highlight.from(token.type()).ordinal()]),
 						token.startOffset(), token.endOffset(), Spanned.SPAN_EXCLUSIVE_EXCLUSIVE);
 			}
 		}
 
+		tokens = newTokens;
+
 		return e;
+	}
+
+	private void provideCompletions(@NonNull List<Token> tokens, @NonNull CharSequence text) {
+		CodeCompletionListener listener = codeCompletionListener;
+		if (listener == null) {
+			return;
+		}
+		int start = getSelectionStart();
+		Token tok = Lexer.findToken(tokens, start);
+		if (tok == null || tok.endOffset() >= text.length()) {
+			listener.onCodeCompletions(DEFAULT_COMPLETIONS, 0);
+			return;
+		}
+		int positionInToken = start - tok.startOffset();
+		listener.onCodeCompletions(
+				Lexer.completeKeyword(
+						Lexer.tokenSource(tok, text).subSequence(0, positionInToken).toString(),
+						tok.category()
+				),
+				positionInToken);
 	}
 
 	private static <T> void clearSpans(Spannable e, int start, int end, Class<T> clazz) {
@@ -593,6 +696,85 @@ public class ShaderEditor extends LineNumberEditText {
 		@Override
 		public void chooseHeight(CharSequence text, int start, int end,
 				int spanstartv, int lineHeight, Paint.FontMetricsInt fm) {
+		}
+	}
+
+	@FunctionalInterface
+	interface OnTokenized {
+		void onTokens(@NonNull List<Token> tokens, @NonNull CharSequence text);
+	}
+
+	static class TokenizeCalculation implements Callable<List<Token>> {
+		private final String text;
+		@NonNull
+		private final OnTokenized onTokenized;
+
+		public TokenizeCalculation(@NonNull String text, @NonNull OnTokenized onTokenized) {
+			this.text = text;
+			this.onTokenized = onTokenized;
+		}
+
+		@Override
+		public List<Token> call() {
+			Lexer lexer = new Lexer(text);
+			List<Token> tokens = new ArrayList<>();
+			for (Token token : lexer) {
+				tokens.add(token);
+				if (Thread.currentThread().isInterrupted()) {
+					break;
+				}
+			}
+			onTokenized.onTokens(tokens, text);
+			return tokens;
+		}
+	}
+
+	static class TokenListUpdater {
+		@NonNull
+		private final OnTokenized onTokenized;
+		@NonNull
+		private final ExecutorService executor = Executors.newSingleThreadExecutor();
+		@Nullable
+		private FutureTask<List<Token>> task;
+		private int revision = -1;
+
+		public TokenListUpdater(@NonNull OnTokenized onTokenized) {
+			this.onTokenized = onTokenized;
+		}
+
+		@NonNull
+		public List<Token> ensureUpdated(@NonNull CharSequence text, int revision) {
+			if (task != null && revision == this.revision) {
+				try {
+					return task.get();
+				} catch (ExecutionException | InterruptedException e) {
+					throw new RuntimeException(e);
+				}
+			}
+			update(text, revision);
+			try {
+				return task.get();
+			} catch (ExecutionException | InterruptedException e) {
+				throw new RuntimeException(e);
+			}
+		}
+
+		public void update(@NonNull CharSequence text, int revision) {
+			if (revision == this.revision) {
+				Log.d("TEST", "Same revision: " + revision);
+				return;
+			} else if (task != null) {
+				task.cancel(false);
+			}
+			Log.d("TEST", "Different revision: " + revision + " != " + this.revision);
+
+			this.revision = revision;
+			task = new FutureTask<>(new TokenizeCalculation(text.toString(), onTokenized));
+			executor.submit(task);
+		}
+
+		public boolean isDone() {
+			return task != null && task.isDone();
 		}
 	}
 }